fix type

+++ webapp/sql/01_schema.sql
@@ -61,6 +61,6 @@ CREATE TABLE `shippings` (
 DROP TABLE IF EXISTS `categories`;
 CREATE TABLE `categories` (
   `id` int unsigned NOT NULL AUTO_INCREMENT PRIMARY KEY,
-  `parent_id` bigint NOT NULL,
+  `parent_id` int unsigned NOT NULL,
   `category_name` varchar(191) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARACTER SET utf8mb4;
